Move metric providers back into language handlers
Reverts part of #2231
