+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/ApexHandler.java
@@ -5,6 +5,8 @@
 package net.sourceforge.pmd.lang.apex;
 
 import java.io.Writer;
+import java.util.Arrays;
+import java.util.List;
 
 import net.sourceforge.pmd.lang.AbstractLanguageVersionHandler;
 import net.sourceforge.pmd.lang.Parser;
@@ -15,11 +17,14 @@ import net.sourceforge.pmd.lang.apex.ast.ASTMethod;
 import net.sourceforge.pmd.lang.apex.ast.ASTUserClassOrInterface;
 import net.sourceforge.pmd.lang.apex.ast.ApexNode;
 import net.sourceforge.pmd.lang.apex.ast.DumpFacade;
-import net.sourceforge.pmd.lang.apex.metrics.ApexMetricsProvider;
+import net.sourceforge.pmd.lang.apex.metrics.ApexMetrics;
+import net.sourceforge.pmd.lang.apex.metrics.api.ApexClassMetricKey;
+import net.sourceforge.pmd.lang.apex.metrics.api.ApexOperationMetricKey;
 import net.sourceforge.pmd.lang.apex.multifile.ApexMultifileVisitorFacade;
 import net.sourceforge.pmd.lang.apex.rule.ApexRuleViolationFactory;
 import net.sourceforge.pmd.lang.ast.xpath.DefaultASTXPathHandler;
 import net.sourceforge.pmd.lang.metrics.LanguageMetricsProvider;
+import net.sourceforge.pmd.lang.metrics.internal.AbstractLanguageMetricsProvider;
 import net.sourceforge.pmd.lang.rule.RuleViolationFactory;
 
 
@@ -64,5 +69,30 @@ public class ApexHandler extends AbstractLanguageVersionHandler {
     @Override
     public LanguageMetricsProvider<ASTUserClassOrInterface<?>, ASTMethod> getLanguageMetricsProvider() {
         return myMetricsProvider;
+    }
+
+    private static class ApexMetricsProvider extends AbstractLanguageMetricsProvider<ASTUserClassOrInterface<?>, ASTMethod> {
+
+        @SuppressWarnings("unchecked")
+        public ApexMetricsProvider() {
+            // a wild double cast
+            super((Class<ASTUserClassOrInterface<?>>) (Object) ASTUserClassOrInterface.class, ASTMethod.class);
+        }
+
+        @Override
+        public List<ApexClassMetricKey> getAvailableTypeMetrics() {
+            return Arrays.asList(ApexClassMetricKey.values());
+        }
+
+
+        @Override
+        protected List<ASTMethod> findOps(ASTUserClassOrInterface<?> astUserClassOrInterface) {
+            return ApexMetrics.findOps(astUserClassOrInterface);
+        }
+
+        @Override
+        public List<ApexOperationMetricKey> getAvailableOperationMetrics() {
+            return Arrays.asList(ApexOperationMetricKey.values());
+        }
     }
 }
