--- pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/metrics/ApexMetricsProvider.java
@@ -1,39 +1,0 @@
-/*
- * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
- */
-
-package net.sourceforge.pmd.lang.apex.metrics;
-
-import java.util.Arrays;
-import java.util.List;
-
-import net.sourceforge.pmd.lang.apex.ast.ASTMethod;
-import net.sourceforge.pmd.lang.apex.ast.ASTUserClassOrInterface;
-import net.sourceforge.pmd.lang.apex.metrics.api.ApexClassMetricKey;
-import net.sourceforge.pmd.lang.apex.metrics.api.ApexOperationMetricKey;
-import net.sourceforge.pmd.lang.metrics.internal.AbstractLanguageMetricsProvider;
-
-public class ApexMetricsProvider extends AbstractLanguageMetricsProvider<ASTUserClassOrInterface<?>, ASTMethod> {
-
-    @SuppressWarnings("unchecked")
-    public ApexMetricsProvider() {
-        // a wild double cast
-        super((Class<ASTUserClassOrInterface<?>>) (Object) ASTUserClassOrInterface.class, ASTMethod.class);
-    }
-
-    @Override
-    public List<ApexClassMetricKey> getAvailableTypeMetrics() {
-        return Arrays.asList(ApexClassMetricKey.values());
-    }
-
-
-    @Override
-    protected List<ASTMethod> findOps(ASTUserClassOrInterface<?> astUserClassOrInterface) {
-        return ApexMetrics.findOps(astUserClassOrInterface);
-    }
-
-    @Override
-    public List<ApexOperationMetricKey> getAvailableOperationMetrics() {
-        return Arrays.asList(ApexOperationMetricKey.values());
-    }
-}
